requirements.txt
pytest~=8.0.0
selene~=2.0.0rc9
selenium~=4.17.2
allure-pytest~=2.13.5
pytest-xdist~=3.5.0
pytest-rerunfailures~=14.0
requests~=2.31.0
pyhamcrest~=2.1.0
Appium-Python-Client~=4.0.0